requirements.txt
chromadb==0.6.3
fastapi==0.115.11
langchain_chroma==0.2.2
langchain_community==0.3.20
langchain_core==0.3.47
langchain_groq==0.3.1
langchain-openai==0.3.9
langchain_huggingface==0.1.2
langchain_text_splitters==0.3.7
langgraph==0.3.18
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.43.2
uvicorn==0.34.0
python-multipart==0.0.20
pypdf==5.4.0
